+++ packages/cms/src/components/fields/DeepSearch.css
@@ -1,0 +1,7 @@
+.cms-deepsearch-container {
+  display: flex;
+}
+
+.cms-deepsearch-list { 
+  outline: 1px solid black;
+}
